Harmonize noodles version across crates, refactor READER/HEADER/INDEX const functions for Search trait(s) (wip). Bump Rust edition to 2021
