+++ htsget-devtools/Cargo.toml
@@ -2,13 +2,11 @@
 name = "htsget-devtools"
 version = "0.1.0"
 authors = ["Christian Perez Llamas <christian.perez@adevinta.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+edition = "2021"
 
 [dependencies]
 
 serde = { version = "~1.0", features = ["derive"] }
 serde_yaml = "~0.8"
 
-noodles = { version = "0.5.0", features = ["bam", "bcf", "bgzf", "cram", "csi", "sam", "tabix", "vcf"] }
+noodles = { version = "0.18.0", features = ["bam", "bcf", "bgzf", "cram", "csi", "sam", "tabix", "vcf"] }
