+++ htsget-http-core/Cargo.toml
@@ -2,7 +2,7 @@
 name = "htsget-http-core"
 version = "0.1.0"
 authors = ["Daniel del Castillo de la Rosa <delcastillodelarosadaniel@gmail.com>", "Marko Malenic <mmalenic1@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [features]
 default = ["async", "htsget-search/default"]
@@ -12,8 +12,8 @@ async = ["tokio", "futures", "htsget-search/async"]
 thiserror = "~1.0"
 serde = { version = "1", features = ["derive"] }
 htsget-search = { path = "../htsget-search" }
-futures = { version = "0.3.16", optional = true }
-tokio = { version = "1.12.0", features = ["full"], optional = true }
+futures = { version = "0.3.21", optional = true }
+tokio = { version = "1.16.1", features = ["full"], optional = true }
 
 [dev-dependencies]
 htsget-id-resolver = { path = "../htsget-id-resolver" }
