+++ htsget-search/Cargo.toml
@@ -2,7 +2,7 @@
 name = "htsget-search"
 version = "0.1.0"
 authors = ["Christian Perez Llamas <chrispz@gmail.com>", "Marko Malenic <mmalenic1@gmail.com>", "Roman Valls Guimera <brainstorm@nopcode.org>"]
-edition = "2018"
+edition = "2021"
 
 [features]
 default = ["async"]
