+++ htsget-search/src/htsget/search.rs
@@ -5,7 +5,6 @@
 //! where the names of the types indicate their purpose.
 //!
 
-use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::Arc;
 
@@ -16,8 +15,10 @@ use noodles::bgzf::VirtualPosition;
 use noodles::csi::binning_index::merge_chunks;
 use noodles::csi::{BinningIndex, BinningIndexReferenceSequence};
 use noodles::sam;
+use noodles_bam::AsyncReader;
 use tokio::fs::File;
 use tokio::io;
+use tokio::io::AsyncRead;
 use tokio::select;
 use tokio::task::JoinHandle;
 
@@ -156,17 +157,21 @@ where
 pub(crate) trait Search<S, ReferenceSequence, Index, Reader, Header>:
   SearchAll<S, ReferenceSequence, Index, Reader, Header>
 where
-  S: AsyncStorage + Send + Sync + 'static,
+  S: AsyncStorage + Send + Sync,
   Header: FromStr + Send,
   Reader: Send,
   Index: Send + Sync,
-  Self: Sync + Send + Sized,
+  Self: Sync + Send,
 {
   const MIN_SEQ_POSITION: u32 = 1; // 1-based
 
-  const READER_FN: fn(File) -> AsyncReader;
-  const HEADER_FN: fn(&'_ mut AsyncReader) -> AsyncHeaderResult;
-  const INDEX_FN: fn(PathBuf) -> AsyncIndexResult<'static, Index>;
+  // const READER_FN: async { fn(AsyncReader<AsyncRead>) -> io::AsyncRead };
+  // const HEADER_FN: async fn(&'_ mut io::AsyncRead) -> AsyncHeaderResult;
+  // const INDEX_FN: async fn(AsyncRead) -> AsyncIndexResult<'static, Index>;
+
+  async fn read(reader: AsyncReader::<dyn AsyncRead>) -> dyn io::AsyncRead;
+  async fn header(header: AsyncReader::<dyn AsyncRead>) -> dyn io::AsyncRead;
+  async fn index(index: AsyncReader::<dyn AsyncRead>) -> dyn io::AsyncRead;
 
   /// Get ranges for a given reference name and an optional sequence range.
   async fn get_byte_ranges_for_reference_name(
